backend/requirements.txt
flask
python-dotenv
flask_sqlalchemy
flask-jwt-extended
flask-cors
Pillow
numpy
requests
